Clean up the actor executor code
